upstream/FANTOM/FANTOM_supplementary/fantom_code/requirement.txt
numpy>=1.19.0
pandas>=1.2.0
torch>=1.10.0
networkx>=2.5
matplotlib>=3.3.0
seaborn>=0.11.0
pyro-ppl>=1.8.0
scikit-learn>=0.24.0
PyYAML>=5.4.1
